refactor: CSS 구조 개선 및 최적화
# 주요 변경사항
1. 폴더 구조 개선
   - styles/
     ├── base/
     ├── components/
     └── layout/

2. CSS 모듈화
   - base/
     ├── reset.css
     ├── typography.css (폰트 및 텍스트 스타일)
     ├── utilities.css (재사용 가능한 유틸리티 클래스)
     └── variables.css (색상 변수)

   - components/
     ├── card.css
     └── column.css

   - layout/
     ├── header.css
     └── kanban.css

3. 컴포넌트 최적화
   - 유틸리티 클래스 적용 (flex-between, flex-center 등)
   - CSS 변수 활용 (색상 시스템)
   - BEM 방식의 클래스 구조화

4. 스타일 시스템 구축
   - 타이포그래피 시스템 정립
   - 색상 시스템 변수화
   - 공통 유틸리티 클래스 정의

# CSS 구조화 및 재사용성 향상을 위한 전반적인 리팩토링
