--- css/icon.css
@@ -1,15 +1,0 @@
-.btn {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 24px;
-    height: 24px;
-    padding: 0;
-    border: none;
-    background-color: transparent;
-    cursor: pointer;
-}
-
-.icon {
-    fill: rgba(160, 163, 189, 1);
-}
