+++ js/main.js
@@ -1,8 +1,5 @@
-import {
-    createCardForm,
-    removeCardForm,
-    createCard
-} from './card-form-template.js';
-import { initModal } from './modal.js';
+import { initCardForm } from './components/CardForm.js';
+import { initModal } from './components/Modal.js';
 
+initCardForm();
 initModal();
