+++ styles/base/utilities.css
@@ -1,0 +1,33 @@
+.flex-between {
+    display: flex;
+    justify-content: space-between;
+    /* align-items: center; */
+}
+
+.flex-row {
+    display: flex;
+    flex-direction: row;
+}
+
+.flex-center {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.flex-column {
+    display: flex;
+    flex-direction: column;
+}
+
+.btn {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 0;
+    border: none;
+    background-color: transparent;
+    cursor: pointer;
+    /* 텍스트 색상 적용 안됨 img 써서*/
+    color: var(--text-weak);
+}
