+++ styles/components/icon-styles.css
@@ -1,0 +1,4 @@
+.small-icon {
+    width: 16px;
+    height: 16px;
+}
