+++ styles/components/modal-styles.css
@@ -24,7 +24,6 @@
 
 .modal__content {
     margin: 0;
-    font-family: Pretendard Variable;
     font-weight: 500;
     font-size: 16px;
     line-height: 22px;
