+++ styles/layout/kanban.css
@@ -1,0 +1,4 @@
+.kanban {
+    display: flex;
+    gap: 24px;
+}
